move system prompts to specific location
