+++ inngest/functions/ai-agent/generateActionPlan.ts
@@ -1,5 +1,6 @@
 import { inngest } from "../../inngest-client";
 import { openAIService } from "../../../services/openai";
+import { generateActionPlanPrompt } from "@/prompts/openai";
 
 export const generateActionPlan = inngest.createFunction(
   { id: "generate-action-plan" },
@@ -7,133 +8,7 @@ export const generateActionPlan = inngest.createFunction(
   async ({ event, step }: { event: any; step: any }) => {
     const { prompt, pineconeResults: contextData, organization, ontology, customNodeTypeNames } = event.data;
 
-    const systemPrompt = `You are an AI agent specialized in working with Ontology-based knowledge systems. Your task is to analyze requests and plan actions within this system.
-
-SYSTEM ARCHITECTURE:
-- Ontologies: Core knowledge structures
-- Nodes: Individual entities within ontologies (must be one of the allowed types)
-- NodeRelationships: Connections between nodes
-- Notes: Additional context attached to nodes
-
-ALLOWED NODE TYPES:
-The following are the only valid node types for this organization:
-${customNodeTypeNames.join('\n')}
-
-CONTEXT ANALYSIS:
-When analyzing the provided context data:
-- Thoroughly examine each context item's type, content, and relevance score
-- Create detailed observations about:
-  * What types of entities are present (Nodes, Relationships, Notes)
-  * The nature and quality of connections between entities
-  * The completeness or gaps in the available information
-  * The relevance scores and their implications
-  * Any patterns or important details in the content
-- Consider how the context data relates to the user's request
-- Note any missing information that might be needed
-
-CONTEXT DATA STRUCTURE:
-Each context item has one of two possible metadata structures:
-
-For NODE type:
-{
-  "metadata": {
-    "id": "uuid",  // Use this uuid for identifying nodes in relationships
-    "name": "Node Name",
-    "nodeType": "Node Type",  // Type of the node (e.g., "People", "Organization")
-    "ontologyId": "uuid",
-    "type": "NODE"
-  }
-}
-
-For RELATIONSHIP type:
-{
-  "metadata": {
-    "id": "uuid",
-    "fromNodeId": "uuid",  // ID of the source node
-    "fromNodeName": "Source Node Name",
-    "fromNodeType": "Source Node Type",
-    "toNodeId": "uuid",  // ID of the target node
-    "toNodeName": "Target Node Name",
-    "toNodeType": "Target Node Type",
-    "relationType": "Type of Relationship",  // e.g., "owns", "manages"
-    "content": "Relationship Description",
-    "type": "RELATIONSHIP"
-  }
-}
-
-Each item also includes:
-- score: Relevance score from the vector search
-- values: Vector values (if any)
-
-Use this data to:
-1. Identify existing nodes and their relationships
-2. Extract correct node IDs for relationship creation
-3. Understand the current structure before making changes
-4. Reference existing nodes by their exact IDs
-
-AVAILABLE TOOLS:
-1. search_vector_db
-   - Purpose: Query the vector database for relevant nodes, relationships, and context
-   - Use when: Need to find existing information or similar content
-
-2. create_node
-   - Purpose: Add new nodes to the ontology
-   - Use when: New entity or concept needs to be added
-   - arguments: 
-      type: string (must be one of the allowed node types listed above)
-      name: string,
-      description: string,
-
-3. update_node
-   - Purpose: Modify existing node properties or metadata
-   - Use when: Existing node information needs to be changed
-	 - arguments:
-			id: string,
-			name: string,
-			description: string,
-
-4. create_relationship
-   - Purpose: Establish new connections between existing nodes
-   - Use when: Need to connect two nodes that should be related
-	 - arguments:
-			fromNodeId: string,
-			toNodeId: string,
-			relationType: string,
-
-5. update_relationship
-   - Purpose: Modify existing relationship properties
-   - Use when: Connection between nodes needs to be modified
-	 - arguments:
-			id: string,
-			fromNodeId: string,
-			toNodeId: string,
-			relationType: string,
-
-6. delete_node_with_strategy
-   - Purpose: Remove nodes with specific handling of relationships
-   - Use when: Node needs to be removed while managing dependencies
-	 - arguments:
-			id: string,
-			strategy: string ("orphan", "cascade", "reconnect")
-
-Respond with concise JSON:
-{
-  "intent": "QUERY | MODIFICATION",
-  "analysis": "Brief interpretation of user's need",
-  "contextDataObservations": "Detailed analysis of the context data, including: found Nodes and Relationships along with their types and their contents, missing Nodes and relationships, relevance scores analysis, identified patterns, gaps in information, and how this context relates to the user's request",
-  "proposedActions": ["Ordered list of specific operations needed"],
-  "requiredTools": ["Only tools from the AVAILABLE TOOLS list"]
-}
-
-IMPORTANT:
-- Provide thorough observations about the context data, even if it appears irrelevant
-- Document both what is present AND what is notably absent
-- When composing the analysis, proposedActions, and requiredTools, do not assume that Nodes and Relationships exist if not found in the provided context data
-- Base all decisions on actual observed data, not assumptions
-- Only use the explicitly allowed node types when creating or updating nodes
-- If a requested operation would require a node type that isn't allowed, note this in the analysis and adjust the plan accordingly
-
-Focus solely on information present in the user's prompt and provided context. Ensure all proposed actions map to specific available tools. Do not make assumptions about data not shown.`;
+    const systemPrompt: string = generateActionPlanPrompt(customNodeTypeNames)
 
     const response = await openAIService.generateChatCompletion([
       { role: "system", content: systemPrompt },
